old-tests/submission/Microsoft/range/range.htm

<!DOCTYPE HTML>
<html>
<head>
  <title>Range Element Tests</title>
  <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
  <link rel="author" title="Microsoft" href="http://www.microsoft.com" />
  <link rel="help" href="http://www.w3.org/TR/html5/Overview.html#range-state" />
</head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div id="log"></div>
<input type="range" id="r00" min="0" max="100" step="20" value="40" style="display:none">
<input type="range" id="r01" min="0" max="1" step=".1" value=".2" style="display:none">
<input type="range" id="r02" style="display:none">

<script>
   function rangeElementTest0() {
       document.getElementById('r00').value = "";
       assert_not_equals(document.getElementById('r00').type, "text");
       assert_not_equals(document.getElementById('r00').value, "");
    }
    test(rangeElementTest0, "range element value set to ''");

   function rangeElementTest1() {
       document.getElementById('r01').value = .6;
       assert_not_equals(document.getElementById('r01').type, "text");
       assert_equals(document.getElementById('r01').value, "0.6");
    }
    test(rangeElementTest1, "range element value set to an integer");

   function rangeElementTest2() {
       assert_not_equals(document.getElementById('r02').type, "text");
       assert_equals(document.getElementById('r02').value, "50");
    }
    test(rangeElementTest2, "range element value equals 50");

   function rangeElementTest3() {
       document.getElementById('r02').value = 200;
       assert_not_equals(document.getElementById('r02').type, "text");
       assert_not_equals(document.getElementById('r02').value, 100);
    }
    test(rangeElementTest3, "range element value equals 100");

   function rangeElementTest4() {
       document.getElementById('r02').value = 2.1;
       assert_not_equals(document.getElementById('r02').type, "text");
       assert_not_equals(document.getElementById('r02').value, 2);
    }
    test(rangeElementTest4, "range element value equals 2");
</script>
</body>
</html>
